System: Add SyncCache
Transforms a thread-unsafe cache into one that can be used by multiple concurrent threads.

`SyncCache::getOrCreate()` will ensure that the cache entry will only be created once even if multiple concurrent calls are made with the same key. Threads waiting for the entry to be created may yield, allowing the thread to do other work instead of blocking until the entry is created.

Bug: b/154138414
Change-Id: I181ebbc3fb01ee62cd8b059fe8ce1c29a6d87865